Re-organise utils test to match utils folder
